+++ common/src/main/java/net/mehvahdjukaar/polytone/texture/DayTimeTexture.java
@@ -17,7 +17,7 @@ public interface DayTimeTexture {
     void polytone$setTimeCycleDuration(int duration);
 
     enum Mode implements StringRepresentable {
-        VANILLA, GAME_TIME, DAY_TIME, WEATHER;
+        VANILLA, GAME_TIME, DAY_TIME, WEATHER, SCREEN_TIME;
 
         @Override
         public String getSerializedName() {