+++ changelog.md
@@ -1,1 +1,2 @@
-- player inventory menu will be refreshed on pack reload so that slot modifiers can be applied properly
+- added screen_time animated texture mode
+- colormap with biome sample can no longer crash. If it were to fail getting a color it will simply output black and log an error. Relevant because of some very rare issue that can happen during a reload on fabric